Ticket #155: trdp_proto.h removed - definitions moved to into iec61375-2-3.h
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@1600 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/example/sendHello.c
@@ -340,7 +340,7 @@ int main (int argc, char *argv[])
         if (outputBuffer != NULL && strlen((char *)outputBuffer) == 0)
         {
             sprintf((char *)outputBuffer, "Just a Counter: %08d", hugeCounter++);
-            outputBufferSize = strlen((char *)outputBuffer);
+            outputBufferSize = (int) strlen((char *)outputBuffer);
         }
 
         err = tlp_put(appHandle, pubHandle, outputBuffer, outputBufferSize);
